requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
cffi==1.14.4
channels==3.0.3
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.5
djangorestframework==3.12.2
ecdsa==0.14.1
hyperlink==21.0.0
idna==3.1
incremental==17.5.0
install==1.3.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==20.0.1
python-jose==3.2.0
pytz==2020.5
rsa==4.7
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
wrapt==1.12.1
zope.interface==5.2.0
